split selection actions into seperate files
still works the exact same, so no problems happen.
but this does make it a bit easier to find where each function is!
